+++ frontend/charts.service.js
@@ -12,7 +12,7 @@ function(rpcService) {
      * Get graph data for specified device
      */
     self.getDeviceData = function(uuid, timestampStart, timestampEnd, options) {
-        return rpcService.sendCommand('get_data', 'charts', {'uuid':uuid, 'timestamp_start':timestampStart, 'timestamp_end':timestampEnd, 'options':options});
+        return rpcService.sendCommand('get_data', 'charts', {'device_uuid':uuid, 'timestamp_start':timestampStart, 'timestamp_end':timestampEnd, 'options':options});
     };
 
 }]);